requirements.txt
Flask
python-dotenv
pdfkit
PyMySQL
cryptography
streamlit
requests
 